requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
dash==0.17.7
dash-core-components==0.12.4
dash-html-components==0.7.0
dash-renderer==0.7.4
decorator==4.1.2
enum34==1.1.6
Flask==0.12.2
Flask-Compress==1.4.0
Flask-SeaSurf==0.2.2
functools32==3.2.3.post2
idna==2.6
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-core==4.3.0
MarkupSafe==1.0
nbformat==4.4.0
nltk==3.2.4
numpy==1.13.1
oauthlib==2.0.2
pandas==0.20.3
pika==0.11.0
plotly==2.0.15
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
textblob==0.13.0
traitlets==4.3.2
tweepy==3.5.0
tweet-preprocessor==0.5.0
urllib3==1.22
Werkzeug==0.12.2
